Send down command ID to datanodes before sending a plan.
Plans are sent down to the datanodes in form of a string and datanodes
convert them back to a RemoteStmt structure using stringToNode() function. Plan
string have relation names which must be converted back to their local OIDs.
For that, datanodes lookup system catalogs. Now if a table is created and later
accessed in the same transaction, and if access requires secondary datanode
connections from oher datanodes, then SnapshotNow fails to find that table.
This happens because command ID is not sent down to the datanode prior to the
plan string. So datanodes use FirstCommandId to do the scan and fails to see
the tuple inserted by the same transaction, but through a diff session.

Fix this my sending down command ID prior to plan. Bug report by Aaron
Jackson and investigation/fix from me.