Fix a bug where a failed PREPARE TRANSACTION would end up rolling back another transaction
This bug was seen when user tries to prepare a transaction with an existing
identifier. In that case, datanodes report error, but the coordinator would
send ROLLBACK PREPARED to those datanodes, thus rolling back already prepared
transaction with the same identifier

Report from Jov, investigations and bug fix from me.